Update deps, rust; address uncovered lints
Update dependencies via `cargo update` and update
rust to the latest (as of this commit) nightly.

Doing this uncovers a few new lints; address those
(they're trivial).  Also, `asm_sym` is now stable
on this nightly, so remove it from the list of
`#![feature(...)]` flags.

+++ src/loader.rs
@@ -106,7 +106,7 @@ fn load_segment(
     section: &ProgramHeader,
     bytes: &[u8],
 ) -> Result<()> {
-    let pa = section.p_paddr as u64;
+    let pa = section.p_paddr;
     if pa % mem::P4KA::ALIGN != 0 {
         return Err("Program section is not physically 4KiB aligned");
     }
